refactor: move langhelpers definitions
